Allow using Unix sockets as locks.
PID file suffer from an issue which is that they are not collected when
the process is abruptly killed. If it happens that before relaunch, a
process reuse the PID of the killed process, then one must externally
clean the PID file.

This patch defines a new compilation flag (--socket) which uses a unix
socket from the abstract namespace as the lock rather than a PID file.

Unfortunately, this option is not portable across all *nix flavours,
only recent Linux versions support it.

A less problematic, but still noticeable effect is that the PID of the
locking process was being advised in the error message. This is not
the case anymore, but this was not the case on Windows anyway.