Refactor and extend benchmarks (see #110)
This adds support for benchmarking multiple acid-state backends
(just the memory and local backends for now). It extends the
benchmarks to measure the initialize operation, and with variants
that issue a query after loading in order to force the state.

The use of criterion has been overhauled to avoid doing unnecessary
work if not all benchmarks are being run.